+++ InnerEye/ML/runner.py
@@ -127,7 +127,7 @@ class Runner:
         # This should be typed as LightningContainer, but we don't always have that imported
         self.lightning_container: Any = None
 
-    def parse_and_load_model(self) -> Optional[ParserResult]:
+    def parse_and_load_model(self) -> ParserResult:
         """
         Parses the command line arguments, and creates configuration objects for the model itself, and for the
         Azure-related parameters. Sets self.azure_config and self.model_config to their proper values. Returns the
@@ -147,7 +147,7 @@ class Runner:
         self.model_config = None
         self.lightning_container = None
         if not azure_config.model:
-            return None
+            raise ValueError("Parameter 'model' needs to be set to tell InnerEye which model to run.")
         model_config_loader: ModelConfigLoader = ModelConfigLoader(**parser_result.args)
         # Create the model as per the "model" commandline option. This can return either a built-in config
         # of type DeepLearningConfig, or a LightningContainer.
@@ -178,7 +178,6 @@ class Runner:
             self.model_config = config_or_container
         else:
             raise ValueError(f"Don't know how to handle a loaded configuration of type {type(config_or_container)}")
-
         if azure_config.extra_code_directory:
             exist = "exists" if Path(azure_config.extra_code_directory).exists() else "does not exist"
             logging.info(f"extra_code_directory is {azure_config.extra_code_directory}, which {exist}")